Removed json Remove json things from ConvertToCSV as that conversion is handled by 'official' tasks. Renamed task to ConvertXmlToCsv.
