+++ Frends.Community.ConvertToCSV.sln
@@ -5,13 +5,13 @@ VisualStudioVersion = 15.0.27130.2024
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{AD988466-462E-48C9-9268-8F420E340A67}"
 	ProjectSection(SolutionItems) = preProject
-		Frends.Community.ConvertToCsv.nuspec = Frends.Community.ConvertToCsv.nuspec
+		Frends.Community.ConvertXmlToCsv.nuspec = Frends.Community.ConvertXmlToCsv.nuspec
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.ConvertToCsv", "Frends.Community.ConvertToCsv\Frends.Community.ConvertToCsv.csproj", "{B9903D7F-C591-463A-94C6-173F1605C22A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.ConvertXmlToCsv", "Frends.Community.ConvertToCsv\Frends.Community.ConvertXmlToCsv.csproj", "{B9903D7F-C591-463A-94C6-173F1605C22A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.ConvertToCsv.Tests", "Frends.Community.ConvertToCsv.Tests\Frends.Community.ConvertToCsv.Tests.csproj", "{17223ED0-3306-4BB2-9E11-108AD44AA58E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.ConvertXmlToCsv.Tests", "Frends.Community.ConvertToCsv.Tests\Frends.Community.ConvertXmlToCsv.Tests.csproj", "{17223ED0-3306-4BB2-9E11-108AD44AA58E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
