+++ README.md
@@ -8,7 +8,7 @@
 
 
 # Frends.Community.ConvertToCsv
-FRENDS Task to convert JSON or XML to CSV. Some Frends4 tasks already provide method .ToCSV() which does the same and this might be obsolete. 
+FRENDS Task to convert XML to CSV. XML needs to represent a table, othervice conversion fails.
 
 ## Contributing
 When contributing to this repository, please first discuss the change you wish to make via issue, email, or any other method with the owners of this repository before making a change.
@@ -27,16 +27,32 @@ NOTE: Be sure to merge the latest from "upstream" before making a pull request!
 
 | Property				|  Type   | Description								| Example                     |
 |-----------------------|---------|-----------------------------------------|-----------------------------|
-| InputData				| string	| Supported formats JSON and XML | `<root><field>1</field></root>` |
-| FileType			| FileType	| JSON or XML	| `Xml` |
-| CsvSeparator			| string	| Separator for the output columns	| `;` |
+| InputData				| string	| XML string to be converted into csv. | `<root>
+	<row id='1'>
+		<name>Google</name>
+		<url>https://en.wikipedia.org/wiki/Google</url>
+		<fancy_characters>comma (,) inside field</fancy_characters>
+	</row>
+	<row id='2'>
+		<name>Apple</name>
+		<url>https://en.wikipedia.org/wiki/Apple_Inc.</url>
+		<fancy_characters>Kanji 漢字</fancy_characters>
+	</row>
+	<row id='3'>
+		<name>Missing columns</name>
+	</row>
+</root>` |
+| CsvSeparator			| string	| Separator for the output columns.	| `;` |
 | IncludeHeaders		| bool	| True if the column headers should be included into the output	| `true` |
 
 ### Result
 
-| Property      | Type     | Description                      |
-|---------------|----------|----------------------------------|
-| Result        | string   | Result as CSV	|
+| Property      | Type     | Description                      | Example                     |
+|---------------|----------|----------------------------------|-----------------------------|
+| Result        | string   | Result as CSV	| `name,url,fancy_characters,id
+Google,https://en.wikipedia.org/wiki/Google,"comma (,) inside field",1
+Apple,https://en.wikipedia.org/wiki/Apple_Inc.,Kanji 漢字,2
+Missing columns,,,3 `
 
 ## License
